+++ app/serializers/daggerheart/character_serializer.rb
@@ -8,7 +8,7 @@ module Daggerheart
                :leveling, :subclasses, :subclasses_mastery, :attacks, :experience, :heritage_name,
                :domains, :selected_domains, :domain_cards_max, :spellcast_traits, :beastform, :beastforms, :tier, :proficiency,
                :can_have_companion, :transformations, :homebrew_domains, :transformation, :can_have_stances, :selected_stances,
-               :stance, :selected_features
+               :stance, :selected_features, :guide_step
 
     delegate :features, :id, :name, :level, :heritage, :main_class, :classes, :traits, :gold, :spent_armor_slots, :health_marked,
              :health_max, :stress_marked, :hope_marked, :stress_max, :hope_max, :modified_traits, :damage_thresholds, :evasion,
@@ -17,7 +17,8 @@ module Daggerheart
              :domain_cards_max, :spellcast_traits, :beastform, :beastforms, :tier, :proficiency,
              :can_have_companion, :transformations, :transformation, :can_have_stances, :selected_stances, :stance,
              :homebrew_domains, :selected_features, to: :decorator
-    delegate :created_at, to: :object
+    delegate :created_at, :data, to: :object
+    delegate :guide_step, to: :data
 
     def provider
       'daggerheart'